==> 01_Scraping_Images.ipynb <==
import numpy as np
import pandas as pd
from bs4 import BeautifulSoup
import time
import re
import os
import urllib.request
from joblib import Parallel, delayed

# pandas options
pd.set_option('display.max_rows', 500)
pd.set_option('display.max_columns', 500)
pd.set_option('display.width', 1000)

# selenium imports
from selenium import webdriver

# selenium preferences to prevent load images
# define webdriver options
chrome_options = webdriver.ChromeOptions()
prefs = {"profile.managed_default_content_settings.images": 2}
chrome_options.add_argument("--headless")
chrome_options.add_argument('window-size=1920x1080');
chrome_options.add_argument("--start-maximized");
chrome_options.add_experimental_option("prefs", prefs)

# read in scraping url
df = pd.read_csv('listing_url.csv')
df.head()

# preprocessing of pandas dataframe
models = []
year = []
listing_id = []

def preprocess_listing_url(x):
    temp = x['listing_url'].split('/')
    listing_id.append(temp[-1])
    
    splits = temp[-2].split('-')
    year.append(splits[0])
    mod_type = ''
    for i in range(1,len(splits)):
        mod_type = mod_type + ' ' + splits[i] 
    models.append(mod_type.strip())
    
df.apply(lambda x: preprocess_listing_url(x), axis = 1)
df['models'] = models
df['year'] = year
df['listing_id'] = listing_id

df.groupby(['models']).agg('count')

# Scale down samples in HRV
df = df[(df.index <= 1230) | (df.index >= 1444)]
df.reset_index(inplace = True)
df.groupby(['models']).agg('count')

df['listing_url'][0]

df['featured'][0]

df['models'][0]

# global vars
filename_list = []
labels_list = []

len(filename_list)

def scrape_images(df, row):
    print('Scraping Row: ', row)
    
    # scrape featured listing => Take only first 4 exterior images
    if df['featured'][row] == True:
        
        # initialize driver
        driver = webdriver.Chrome(options=chrome_options)
        driver.get(df['listing_url'][row])
        
        currlisting = df['listing_id'][row]
        im_count = 0
        
        # find image xpath
        for i in range(4):
            # find and save images with unique filename (listing_id_im_count.jpeg)
            img = driver.find_element_by_xpath("//*[@id='root']/div[1]/main/div/div[1]/div/div[1]/div/div[" + str(i+1) + "]/a/div")
            img = img.find_element_by_tag_name("img")
            img_url = img.get_property('src')
            extension = img_url.split('.')[-1]
            
            filename = currlisting + '_' + str(im_count) + '.' + extension
            labels = df['models'][row]
            
            urllib.request.urlretrieve(img_url, os.path.join(os.getcwd(), 'dataset', filename))
            im_count = im_count + 1
            
            # book keep labels and filenames
            filename_list.append(filename)
            labels_list.append(labels)
            
            # click next button
            next_button = driver.find_element_by_class_name('SlideshowArrow-Right')
            next_button.click()
        
        # close driver
        driver.close()

    # non featured listing
    else:
        
        # initialize driver
        driver = webdriver.Chrome(options=chrome_options)
        driver.get(df['listing_url'][row])
        
        currlisting = df['listing_id'][row]
        im_count = 0
        
        cont = driver.find_element_by_class_name('SlideshowChip_Container')
        capt = cont.find_elements_by_class_name('fg-white')
        num_images = int(capt[1].text.split('/')[-1].strip())
        
        # find image xpath
        for i in range(num_images):
            # find and save images with unique filename (listing_id_im_count.jpeg)
            img = driver.find_element_by_xpath("//*[@id='root']/div[1]/main/div/div[1]/div/div[1]/div/div[" + str(i+1) + "]/a/div")
            img = img.find_element_by_tag_name("img")
            img_url = img.get_property('src')
            extension = img_url.split('.')[-1]
            
            filename = currlisting + '_' + str(im_count) + '.' + extension
            labels = df['models'][row]
            
            urllib.request.urlretrieve(img_url, os.path.join(os.getcwd(), 'dataset', filename))
            im_count = im_count + 1
            
            # book keep labels and filenames
            filename_list.append(filename)
            labels_list.append(labels)
            
            # click next buttion
            next_button = driver.find_element_by_class_name('SlideshowArrow-Right')
            next_button.click()
        
        # close driver
        driver.close()
        
        time.sleep(4)

Parallel(n_jobs=-1, require='sharedmem')(delayed(scrape_images)(df, i) for i in range(2336, df.shape[0]));

if len(np.unique(filename_list)) == len(os.listdir(os.path.join(os.getcwd(), 'dataset'))):
    print('Same length can continue')
else:
    print('different length!, please investigate')

# stroe all labels
labels_df = pd.DataFrame()
labels_df['file_name'] = filename_list
labels_df['labels'] = labels_list 
labels_df.head()

# save to output
labels_df.to_csv('labels_df.csv', index = False)

labels_df.shape

len(os.listdir(os.path.join(os.getcwd(), 'dataset')))
