==> 02_Train_test_split.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import os
import random
from sklearn.model_selection import train_test_split
import shutil

data_path = os.path.join(os.getcwd(), 'dataset')

file_list = os.listdir(data_path)
df = pd.read_csv('labels_df.csv')

df.head()

df = df[df['file_name'].isin(file_list)]
print('Filtered_df shape: ', df.shape)
print('Filelist length: ', len(file_list))

X = np.array(df['file_name'])
y = np.array(df['labels'])

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)
X_train, X_val, y_train, v_val = train_test_split(X_train, y_train, test_size=0.05, random_state = 42)

print('Train size: ', X_train.shape[0])
print('Val size: ', X_val.shape[0])
print('Test size: ', X_test.shape[0])

# move files to destination
for tr in X_train:
    # Move a file from the directory d1 to d2
    shutil.move(os.path.join(os.getcwd(), 'dataset', tr), os.path.join(os.getcwd(), 'dataset_final', 'train', tr))

for v in X_val:
    # Move a file from the directory d1 to d2
    shutil.move(os.path.join(os.getcwd(), 'dataset', v), os.path.join(os.getcwd(), 'dataset_final', 'val', v))

for t in X_test:
    # Move a file from the directory d1 to d2
    shutil.move(os.path.join(os.getcwd(), 'dataset', t), os.path.join(os.getcwd(), 'dataset_final', 'test', t))

# create labels in csv format
train_df = df[df['file_name'].isin(X_train)]
val_df = df[df['file_name'].isin(X_val)]
test_df = df[df['file_name'].isin(X_test)]

train_df.to_csv('labels_train.csv', index = False)
val_df.to_csv('labels_val.csv', index = False)
test_df.to_csv('labels_test.csv', index = False)
